feat(useEndpointForVADCDataDictionary): Update API Call to use Fetch instead of JSON file import
