+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/AtlasDataDictionaryTable.test.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import AtlasDataDictionaryContainer from './AtlasDataDictionaryContainer';
+import AtlasDataDictionaryTable from './AtlasDataDictionaryTable';
+import TableData from './TestData/TableData';
 
 const atlasDataDictionaryUITestIds = [
-  'atlas-data-dictionary-container',
+  'atlas-data-dictionary-table',
   'search-bar',
   'column-headers',
   'entries-header',
@@ -13,7 +14,7 @@ const atlasDataDictionaryUITestIds = [
 
 describe('AtlasDataDictionaryContainer', () => {
   it('renders the AtlasDataDictionaryContainer component and associated UI elements correctly', () => {
-    render(<AtlasDataDictionaryContainer />);
+    render(<AtlasDataDictionaryTable TableData={TableData} />);
     atlasDataDictionaryUITestIds.forEach((testID) => {
       expect(screen.getByTestId(testID)).toBeInTheDocument();
     });
