+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Components/TableRow/NonNumericDetailsTable/NonNumericDetailsTable.test.tsx
