+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Components/TableRow/NumericDetailsTable/NumericDetailsTable.test.tsx
