+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Components/TableRow/NumericDetailsTable/NumericDetailsTable.tsx
