+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Interfaces/Interfaces.ts
