+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Utils/DetermineInitialAtlasColumnManagement.ts
